# OpenAPI/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/users:
    post:
      tags:
      - Users
      summary: Create new user
      description: Create a new user account
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUserDTO"
        required: true
      responses:
        "200":
          description: User created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags:
      - Users
      summary: Delete current user
      description: Delete the authenticated user's account
      operationId: deleteCurrentlyAuthenticatedUser
      responses:
        "401":
          description: Unauthorized - authentication required
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
        "204":
          description: User deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /api/users/verify-password:
    post:
      tags:
      - Users
      summary: Verify password
      description: Verify if the provided password matches the authenticated user's
        password
      operationId: verifyPassword
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: Password verification result returned
          content:
            '*/*':
              schema:
                type: boolean
        "401":
          description: Unauthorized - authentication required
          content:
            '*/*':
              schema:
                type: boolean
  /api/users/password:
    post:
      tags:
      - Users
      summary: Change password
      description: Change the authenticated user's password
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "401":
          description: Unauthorized - authentication required
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Password changed successfully
          content:
            '*/*':
              schema:
                type: string
...
  /api/users/{id}:
    get:
      tags:
      - Users
      summary: Get user by ID
      description: Retrieve a user's information by their ID
      operationId: getUser
      parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    patch:
      tags:
      - Users
      summary: Update user
      description: Partially update a user's information
      operationId: patchUser
      parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: User updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
...
  /api/users/me:
    get:
      tags:
      - Users
      summary: Get current user
      description: Retrieve the authenticated user's information
      operationId: getCurrentUser
      responses:
        "401":
          description: Unauthorized - authentication required
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
        "200":
          description: Current user information retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
  /api/users/IsUsernameTaken:
    get:
      tags:
      - Users
      summary: Check username availability
      description: Check if a username is already taken
      operationId: isUsernameTaken
      parameters:
      - name: username
        in: query
        description: Username to check
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Username availability checked
          content:
            '*/*':
              schema:
                type: boolean
...
  /api/murals/{muralId}/users/{userId}:
    delete:
      tags:
      - Murals
      summary: Remove user from mural
      description: Remove a specified user from a mural
      operationId: removeSpecifiedUserFromMural
      parameters:
      - name: muralId
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: User ID to remove
        required: true
        schema:
          type: integer
          format: int64
      - name: inheritor
        in: query
        description: New owner ID if current owner is leaving
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural or user not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "200":
          description: User removed from mural
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "403":
          description: Forbidden - insufficient permissions
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
  /api/murals/{id}/users/me:
    delete:
      tags:
      - Murals
      summary: Leave mural
      description: Remove the current user from a mural
      operationId: removeCurrentUserFromMural
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      - name: inheritor
        in: query
        description: New owner ID if current owner is leaving
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "200":
          description: User removed from mural
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
...
components:
  schemas:
    NewUserDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: string
    Mural:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        code:
          type: string
        visibility:
          type: string
          enum:
          - PRIVATE
          - MURAL_SPECIFIC
          - MURAL_PUBLIC
          - PUBLIC
        owner:
          $ref: "#/components/schemas/User"
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        bannedUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        thumbnail:
          type: array
          items:
            type: string
            format: byte
        banner:
          type: array
          items:
            type: string
            format: byte
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: string
        memberMurals:
          type: array
          items:
            $ref: "#/components/schemas/Mural"
        roles:
          type: array
          items:
            type: string
        admin:
          type: boolean
...
    NamedId:
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int64
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
...